+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/spec/dsl/aftereach/AfterSpecTest.kt
@@ -1,17 +1,16 @@
-package com.sksamuel.kotest.listeners.testlistener.instancepertest
+package com.sksamuel.kotest.engine.spec.dsl.aftereach
 
 import io.kotest.core.listeners.TestListener
 import io.kotest.core.spec.IsolationMode
+import io.kotest.core.spec.Spec
 import io.kotest.core.spec.style.FunSpec
-import io.kotest.core.test.TestCase
-import io.kotest.core.test.TestResult
 import io.kotest.matchers.shouldBe
 import java.util.concurrent.atomic.AtomicInteger
 
-class AfterTestTest : FunSpec() {
+class AfterSpecTest : FunSpec() {
 
    val listener = object : TestListener {
-      override suspend fun afterTest(testCase: TestCase, result: TestResult) {
+      override suspend fun afterSpec(spec: Spec) {
          counter.incrementAndGet()
       }
    }
@@ -27,7 +26,7 @@ class AfterTestTest : FunSpec() {
       register(listener)
 
       afterProject {
-         counter.get() shouldBe 4
+         counter.get() shouldBe 5
       }
 
       test("ignored test").config(enabled = false) {}
