+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/spec/dsl/aftereach/BeforeTestTest.kt
@@ -1,16 +1,16 @@
-package com.sksamuel.kotest.listeners.testlistener.instancepertest
+package com.sksamuel.kotest.engine.spec.dsl.aftereach
 
 import io.kotest.core.listeners.TestListener
 import io.kotest.core.spec.IsolationMode
-import io.kotest.core.spec.Spec
 import io.kotest.core.spec.style.FunSpec
+import io.kotest.core.test.TestCase
 import io.kotest.matchers.shouldBe
 import java.util.concurrent.atomic.AtomicInteger
 
-class BeforeSpecTest : FunSpec() {
+class BeforeTestTest : FunSpec() {
 
    val listener = object : TestListener {
-      override suspend fun beforeSpec(spec: Spec) {
+      override suspend fun beforeTest(testCase: TestCase) {
          counter.incrementAndGet()
       }
    }
@@ -23,15 +23,13 @@ class BeforeSpecTest : FunSpec() {
 
    init {
 
-      register(listener)
+      listener(listener)
 
       afterProject {
          counter.get() shouldBe 4
       }
 
-      // this shouldn't trigger the after spec as its in an isolated instance
       test("ignored test").config(enabled = false) {}
-
       test("a").config(enabled = true) {}
       test("b").config(enabled = true) {}
       test("c").config(enabled = true) {}
